🩹 Refactor : useProductRegistration 훅을 useProduct.js로 이동 및 불필요한 파일 삭제 #68
